moved code to andrewoid/ and added comment to Digit.getLinesStatus()
