+++ example/src/main/java/andrewoid/neuralnetwork/DigitInteraction.java
@@ -1,7 +1,8 @@
-package ayeletbuchen;
+package andrewoid.neuralnetwork;
 
 import andrewoid.neutralnetwork.Network;
 import andrewoid.neutralnetwork.Neuron;
+import andrewoid.neutralnetwork.DigitFrame;
 
 import java.util.Random;
