+++ library/src/main/java/andrewoid/neutralnetwork/EvaluateButton.java
@@ -1,7 +1,5 @@
-package ayeletbuchen;
+package andrewoid.neutralnetwork;
 
-import andrewoid.neutralnetwork.Network;
-import andrewoid.neutralnetwork.Neuron;
 import javax.swing.*;
 
 public class EvaluateButton extends JButton {
